requirements/conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.11.0=pypi_0
appdirs=1.4.4=py_0
argon2-cffi=20.1.0=py38h7b6447c_1
ase=3.20.1=pypi_0
async_generator=1.10=py_0
attrs=20.2.0=py_0
backcall=0.2.0=py_0
black=19.10b0=py_0
blas=1.0=mkl
bleach=3.2.1=py_0
ca-certificates=2021.1.19=h06a4308_1
cached-property=1.5.2=pypi_0
cachetools=4.1.1=pypi_0
certifi=2020.12.5=py38h06a4308_0
cffi=1.14.3=py38he30daa8_0
chardet=3.0.4=pypi_0
click=7.1.2=pyhd3eb1b0_0
configparser=5.0.1=pypi_0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.10.0=pypi_0
dataclasses=0.6=pypi_0
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
docker-pycreds=0.4.0=pypi_0
entrypoints=0.3=py38_0
expat=2.2.10=he6710b0_2
fairlearn=0.5.0=pypi_0
flake8=3.9.0=pyhd3eb1b0_0
fontconfig=2.13.0=h9420a91_0
freetype=2.10.4=h5ab3b9f_0
future=0.18.2=pypi_0
gensim=3.8.3=pypi_0
gitdb=4.0.5=pypi_0
gitpython=3.1.12=pypi_0
glib=2.66.1=h92f7085_0
google-auth=1.23.0=pypi_0
google-auth-oauthlib=0.4.2=pypi_0
googledrivedownloader=0.4=pypi_0
grpcio=1.33.2=pypi_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb31296c_0
h5py=3.0.0=pypi_0
icu=58.2=he6710b0_3
idna=2.10=pypi_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
intel-openmp=2020.2=254
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.19.0=py38hb070fc8_0
ipython_genutils=0.2.0=py38_0
ipywidgets=7.5.1=py_1
isodate=0.6.0=pypi_0
jedi=0.17.2=py38_0
jinja2=2.11.2=py_0
joblib=0.17.0=pypi_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=py_2
jupyter=1.0.0=py38_7
jupyter_client=6.1.7=py_0
jupyter_console=6.2.0=py_0
jupyter_core=4.6.3=py38_0
jupyterlab_pygments=0.1.2=py_0
kiwisolver=1.3.1=pypi_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
llvmlite=0.34.0=pypi_0
lz4-c=1.9.2=heb0550a_3
markdown=3.3.3=pypi_0
markupsafe=1.1.1=py38h7b6447c_0
matplotlib=3.3.2=pypi_0
mccabe=0.6.1=py38_1
mistune=0.8.4=py38h7b6447c_1000
mkl=2020.2=256
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.2.0=py38h23d657b_0
mkl_random=1.1.1=py38h0573a6f_0
mypy_extensions=0.4.3=py38_0
natsort=7.1.0=pypi_0
nbclient=0.5.1=py_0
nbconvert=6.0.7=py38_0
nbformat=5.0.8=py_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.4.1=py_0
networkx=2.5=pypi_0
ninja=1.10.1=py38hfd86e86_0
notebook=6.1.4=py38_0
numba=0.51.2=pypi_0
numpy=1.19.2=py38h54aff64_0
numpy-base=1.19.2=py38hfa32c7d_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py_0
opencv-python=4.5.1.48=pypi_0
openssl=1.1.1k=h27cfd23_0
packaging=20.4=py_0
pandas=1.1.4=pypi_0
pandoc=2.11=hb0f4dca_0
pandocfilters=1.4.2=py38_1
parso=0.7.0=py_0
pathspec=0.7.0=py_0
pathtools=0.1.2=pypi_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=py38_0
pickleshare=0.7.5=py38_1000
pillow=8.0.1=py38he98fc37_0
pip=20.2.4=py38h06a4308_0
prometheus_client=0.8.0=py_0
promise=2.3=pypi_0
prompt-toolkit=3.0.8=py_0
prompt_toolkit=3.0.8=0
protobuf=3.13.0=pypi_0
psutil=5.8.0=pypi_0
ptyprocess=0.6.0=py38_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycodestyle=2.6.0=py_0
pycparser=2.20=py_2
pyflakes=2.2.0=py_0
pygments=2.7.2=pyhd3eb1b0_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py38h05f1152_4
pyrsistent=0.17.3=py38h7b6447c_0
python=3.8.5=h7579374_1
python-dateutil=2.8.1=py_0
python-louvain=0.15=pypi_0
pytorch=1.7.0=py3.8_cuda10.1.243_cudnn7.6.3_0
pytz=2020.4=pypi_0
pyyaml=5.3.1=pypi_0
pyzmq=19.0.2=py38he6710b0_1
qt=5.9.7=h5867ecd_1
qtconsole=4.7.7=py_0
qtpy=1.9.0=py_0
rdflib=5.0.0=pypi_0
readline=8.0=h7b6447c_0
regex=2020.11.13=py38h27cfd23_0
requests=2.24.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.6=pypi_0
scikit-learn=0.23.2=pypi_0
scipy=1.5.3=pypi_0
seaborn=0.11.1=pypi_0
send2trash=1.5.0=py38_0
sentry-sdk=0.19.5=pypi_0
setuptools=50.3.0=py38h06a4308_1
shortuuid=1.0.1=pypi_0
sip=4.19.13=py38he6710b0_0
six=1.15.0=py_0
smart-open=4.0.0=pypi_0
smmap=3.0.4=pypi_0
sqlite=3.33.0=h62c20be_0
subprocess32=3.5.4=pypi_0
tensorboard=2.3.0=pypi_0
tensorboard-plugin-wit=1.7.0=pypi_0
tensorboardx=2.1=pypi_0
terminado=0.9.1=py38_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=hbc83047_0
toml=0.10.1=py_0
torch-cluster=1.5.8=pypi_0
torch-geometric=1.6.3=pypi_0
torch-scatter=2.0.5=pypi_0
torch-sparse=0.6.8=pypi_0
torch-spline-conv=1.2.0=pypi_0
torchaudio=0.7.0=py38
torchvision=0.8.1=py38_cu101
tornado=6.0.4=py38h7b6447c_1
tqdm=4.41.1=pypi_0
traitlets=5.0.5=py_0
typed-ast=1.4.2=py38h27cfd23_1
typing_extensions=3.7.4.3=py_0
urllib3=1.25.11=pypi_0
wandb=0.10.13=pypi_0
watchdog=0.10.4=pypi_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py38_1
werkzeug=1.0.1=pypi_0
wheel=0.35.1=py_0
widgetsnbextension=3.5.1=py38_0
xz=5.2.5=h7b6447c_0
zeromq=4.3.3=he6710b0_3
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
